Change module style name (#16)
This is because the theme used in our docs seems to have very slight
differences. Before this change, modules had the same color as operators
and functions. Now, they have a different color (orange).

Could have also changed the docs theme instead, but I'm not sure who has
the "right" color. Thus I opted to choose a class where both themes use
the same color. Note that this class might be rendered differently in
non-GitHub themes than before. But for our purposes this does the job.